CDMS-366: code review change
