+++ src/plugins/auth/defra-id.js
@@ -1,8 +1,7 @@
 import bell from '@hapi/bell'
 import { paths } from '../../routes/route-constants.js'
 import { config, configKeys } from '../../config/config.js'
-import { getDefraIdAuthConfig } from '../../auth/defraId-client.js'
-import { supplyProfileFunc } from './utils/profile-provider.js'
+import { defraIdAuthProvider } from '../../auth/defra-id-auth-provider.js'
 
 const authConfig = config.get('auth.defraId')
 const sessionConfig = config.get('session')
@@ -11,7 +10,6 @@ const defraId = {
   plugin: {
     name: 'defra-id',
     register: async (server) => {
-      const oidcConfigurationUrl = authConfig.oidcConfigurationUrl
       const serviceId = authConfig.serviceId
       const clientId = authConfig.clientId
       const clientSecret = authConfig.clientSecret
@@ -19,24 +17,13 @@ const defraId = {
 
       await server.register(bell)
 
-      const oidcConf = await getDefraIdAuthConfig(oidcConfigurationUrl)
-
       server.auth.strategy('defra-id', 'bell', {
         location: (request) => {
           request.yar.flash('referrer', paths.SEARCH)
 
           return authCallbackUrl
         },
-        provider: {
-          name: 'defra-id',
-          protocol: 'oauth2',
-          useParamsAuth: true,
-          auth: oidcConf.authorization_endpoint,
-          token: oidcConf.token_endpoint,
-          pkce: 'S256',
-          scope: authConfig.scopes,
-          profile: await supplyProfileFunc(oidcConf)
-        },
+        provider: await defraIdAuthProvider(),
         password: sessionConfig.cookie.password,
         clientId,
         clientSecret,
